Generalized six wheel tank drive code and added to Robotnik
Each year, similar code is used for drive bases.
However, it remains in the 'per year' code, as if it always changes.
This commit is to provide a format for generalizing drive base code, to
simplify code reuse each year.

The gyro has been removed, as it never seemed to be nessessary in the
drive base itself.